requirements.txt
python-dotenv>=1.0.0
# Agentrics SDK — install from the local repo (adjust path if needed)
-e ../agent-cost-optimizer/sdk/python
